requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
azure-eventhub==5.14.0
python-dotenv==1.0.1
faker==33.3.0